+++ CHANGELOG.md
@@ -1,4 +1,25 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2026-07-25 05:11:41
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 092f047adfb35aced830f0dba436a968bfaf0fdf
+
+## [ForwardWidgets-2kuai] - 2026-07-25 05:11:41
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: e5003ef68c04125c21ac490976d8a0b04196f8db
+
+## [Forward-xbzl] - 2026-07-25 05:11:41
+- 仓库: https://github.com/xbzl/Forward
+- 分支: main
+- 新增: dfd071a87b7f3b02d2aefec9fbae9ca072d32c33
+
+## [scripts-bemarkt] - 2026-07-25 05:11:41
+- 仓库: https://github.com/bemarkt/scripts
+- 分支: master
+- 新增: 2b765c63b50866444382e0fd3b92683badd1b7b9
+
+
 ## [ForwardWidgets-huangxd] - 2026-07-24 05:16:55
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main